Integrate PaymentFlowResultProcessor in DefaultFlowController
Convert `PaymentFlowResultProcessor` to an interface and move
implementation to `DefaultPaymentFlowResultProcessor`.

This allows `DefaultFlowController` to handle payment authentication
results.
